#22 presentationSummary
