+++ server/src/models/versionModels.ts
@@ -1,0 +1,5 @@
+export interface IVersion {
+  _id: number
+  number: number
+  slides: string[]
+}
